# 4059 ensures substitution