--- a/booster/library/Booster/Log.hs
+++ b/booster/library/Booster/Log.hs
@@ -68,6 +68,7 @@ import Booster.Pattern.Base (
     TermAttributes (hash),
     pattern AndTerm,
  )
+import Booster.Pattern.Bool (asEquations)
 import Booster.Pattern.Pretty
 import Booster.Prettyprinter (renderOneLineText)
 import Booster.Syntax.Json (KorePattern, addHeader, prettyPattern)
@@ -185,8 +186,8 @@ withTermContext t@(Term attrs _) m =
         m
 
 withPatternContext :: LoggerMIO m => Pattern -> m a -> m a
-withPatternContext Pattern{term, constraints} m =
-    let t' = foldl' AndTerm term $ Set.toList $ Set.map coerce constraints -- FIXME
+withPatternContext Pattern{term, constraints, substitution} m =
+    let t' = foldl' AndTerm term . map coerce $ Set.toList constraints <> asEquations substitution
      in withTermContext t' m
 
 instance ToLogFormat KorePattern where