--- a/booster/library/Booster/Syntax/Json/Internalise.hs
+++ b/booster/library/Booster/Syntax/Json/Internalise.hs
@@ -24,7 +24,6 @@ module Booster.Syntax.Json.Internalise (
     textToBS,
     trm,
     handleBS,
-    asEquations,
     TermOrPredicates (..),
     InternalisedPredicates (..),
     PatternOrTopOrBottom (..),