@geo2a 2024-10-17
Failed KEVM tests as of [52d3ad4](https://github.com/runtimeverification/haskell-backend/pull/4060/commits/52d3ad480c0390a131446c51a3480427d2299779)

```
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/flopper-dent-guy-diff-tic-not-0-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/flopper-dent-guy-same-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/flopper-kick-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/cat-file-addr-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/pot-join-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/end-pack-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/flapper-yank-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_prove_rules[mcd/flopper-cage-pass-spec.k]
```