@geo2a 2024-10-22
No failed KEVM tests at commit https://github.com/runtimeverification/haskell-backend/commit/fdf62fc063fe4a8379efc05d74614fc83725bcfc and not much difference in performance:
```
| Test                                          | 4059-ensures-substitution time | master-434710529 time | (4059-ensures-substitution/master-434710529) time
|-----------------------------------------------|--------------------------------|-----------------------|---------------------------------------------------
| mcd/vow-fess-fail-rough-spec.k                | 108.68                         | 115.11                | 0.9441403874554775
| mcd-structured/flopper-file-pass-rough-spec.k | 18.23                          | 19.14                 | 0.9524555903866249
| erc20/ds/allowance-spec.k                     | 52.94                          | 55.03                 | 0.9620207159731056
| benchmarks/storagevar01-spec.k                | 49.16                          | 47.2                  | 1.0415254237288134
| mcd/flopper-file-pass-rough-spec.k            | 19.46                          | 18.58                 | 1.047362755651238
| mcd-structured/cat-exhaustiveness-spec.k      | 118.05                         | 111.06                | 1.0629389519178822
| TOTAL                                         | 366.52000000000004             | 366.12                | 1.0010925379656943
```